feat : 리뷰 반영 except 게터 세터
