+++ src/main/java/codesquad/QuestionRepository.java
@@ -1,6 +1,8 @@
 package codesquad;
 
+import codesquad.domain.Question;
 import org.springframework.data.repository.CrudRepository;
 
 public interface QuestionRepository extends CrudRepository<Question, Long> {
+
 }
