--- src/main/java/codesquad/User.java
@@ -1,84 +1,0 @@
-package codesquad;
-
-import lombok.*;
-import javax.persistence.*;
-
-@Getter
-@Setter
-@ToString
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 20)
-    private String userId;
-
-    private String password;
-    private String name;
-    private String email;
-    private String changePassword;
-
-    /*
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getChangePassword() {
-        return changePassword;
-    }
-
-    public void setChangePassword(String changePassword) {
-        this.changePassword = changePassword;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", userId='" + userId + '\'' +
-                ", password='" + password + '\'' +
-                ", name='" + name + '\'' +
-                ", email='" + email + '\'' +
-                ", changePassword='" + changePassword + '\'' +
-                '}';
-    }
-    */
-}
