+++ src/main/java/codesquad/UserRepository.java
@@ -1,5 +1,6 @@
 package codesquad;
 
+import codesquad.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Long> {
