+++ src/main/java/codesquad/controller/UserController.java
@@ -1,5 +1,7 @@
-package codesquad;
+package codesquad.controller;
 
+import codesquad.UserRepository;
+import codesquad.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,7 +9,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
 
 @Controller
 @RequestMapping("/users")
@@ -30,23 +31,22 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ModelAndView show(@PathVariable("id") Long id) {
-        ModelAndView mav = new ModelAndView("user/profile");
-        mav.addObject("user", userRepository.findById(id).get());
-        return mav;
+    public String show(@PathVariable("id") Long id, Model model) {
+        model.addAttribute("user", userRepository.findById(id).get());
+        return "user/profile";
     }
 
     @GetMapping("/{id}/form")
-    public ModelAndView updateForm(@PathVariable("id") Long id) {
-        ModelAndView mav = new ModelAndView("user/updateForm");
-        mav.addObject("user", userRepository.findById(id).get());
-        return mav;
+    public String updateForm(@PathVariable("id") Long id, Model model) {
+        model.addAttribute("user", userRepository.findById(id).get());
+        return "user/updateForm";
     }
 
+    //setter사용금지
     @PostMapping("/{id}/update")
     public String editUser(@PathVariable("id") Long id, User user) {
         User changedUser = userRepository.findById(id).get();
-        if (changedUser.getPassword().equals(user.getPassword())) {
+        if (changedUser.isSamePassword(changedUser,user)) {
             changedUser.setName(user.getName());
             changedUser.setEmail(user.getEmail());
             changedUser.setPassword(user.getChangePassword());
