+++ src/main/java/codesquad/domain/Question.java
@@ -1,0 +1,31 @@
+package codesquad.domain;
+
+import javax.persistence.*;
+import lombok.*;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+@Getter
+@Setter
+@ToString
+@Entity
+public class Question {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 20)
+    private String writer;
+
+    private String title;
+    private String contents;
+    private String time;
+
+    public void checkCurrentTime(){
+        Date today = new Date();
+        SimpleDateFormat date = new SimpleDateFormat("yyyy-MM-dd");
+        SimpleDateFormat time = new SimpleDateFormat("HH:mm");
+        this.time = date.format(today) + " " + time.format(today);
+    }
+}
